Separate sagas to entities
