+++ src/sagas/fetchData.js
@@ -1,0 +1,4 @@
+export default async function fetchData(url) {
+  const response = await fetch(url);
+  return await response.json();
+}
