+++ src/sagas/index.js
@@ -1,7 +1,9 @@
 import { all } from "redux-saga/effects";
 
-import { charactersWatcher } from "./data";
+import { charactersWatcher } from "./characters";
+import { locationsWatcher } from "./locations";
+import { episodesWatcher } from "./episodes";
 
 export default function* rootSaga() {
-  yield all([charactersWatcher()]);
+  yield all([charactersWatcher(), locationsWatcher(), episodesWatcher()]);
 }
